+++ src/main/kotlin/README.md
@@ -7,7 +7,7 @@
 - 카드를 안받는다는 의사를 밝힐 때까지 계속 받을 수 있음 (단 합이 21이 넘으면 자동 패배)
 
 ### TODO
-- [ ] 카드 덱 구성
+- [X] 카드 덱 구성
   - Ace, 1 - 10, King, Queen, Jack | 다이아, 클로버, 하트, 스페이드
   - 무늬는 게임에 영향을 주지 않음
   - 숫자 하나당 4 종류의 카드
